test: consolidate load-balancer feedback coverage
